efi: Add grub_efi_allocate_pages_max() and grub_efi_allocate_aligned_max()
This patch adds grub_efi_allocate_pages_max() on UEFI systems, which
is used to request a memory allocation with a limit on the maximum
address that will be used.

It also adds grub_efi_allocate_aligned_max(), which allocates pages
under an address limit, aligned to a specific alignment.

Signed-off-by: Peter Jones <pjones@redhat.com>