types.h: Fix GRUB_SIZE_MAX on platforms where long is 32-bits.
On LLP64 platforms, "unsigned long long" and pointers are 64-bit, but
"unsigned long" is 32-bit.  At the same time, on LP64 platforms all of
these are 64-bit.  Which means currently on LLP64 systems, GRUB_SIZE_MAX
is a literal with a value that is too large for the type it evaluates as
(UL).

Since ULL is 64-bits in all cases, just use that instead.

Signed-off-by: Peter Jones <pjones@redhat.com>